requirements.txt
Flask
requests
newsapi-python
g4f